refactor(slider-range): rename track to active-track component
