+++ src/components/common/slider-range/slider-range-active-track.vue
